+++ app/components/directives/grid/grid.js
@@ -1,0 +1,13 @@
+/**
+ * Created by pua on 9/17/2016.
+ */
+/**
+ * Created by pua on 9/15/2016.
+ */
+angular.module("iAngelsTeaShop")
+    .directive("gridDirective", function() {
+        return {
+            templateUrl : "app/components/directives/grid/grid.html",
+        };
+    });
+