+++ app/components/directives/gridTile/gridTile.js
@@ -1,0 +1,10 @@
+/**
+ * Created by pua on 9/15/2016.
+ */
+angular.module("iAngelsTeaShop")
+    .directive("gridTileDirective", function() {
+        return {
+            templateUrl : "app/components/directives/gridTile/gridTile.html",
+        };
+    });
+