+++ app/config.js
@@ -1,0 +1,8 @@
+/**
+ * Created by pua on 9/17/2016.
+ */
+angular.module('iAngelsTeaShop').constant('config',
+    {
+        mock: 'true'
+    }
+);