+++ Changes.md
@@ -7,10 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Current
 ### Added
 
-Added Cart page template with list and checkout button
-Added angular material dependency
+
 
 ### Changed
+
+## [0.0.0] - 2015-09-17
+### Added
+- Added Cart page template with list and checkout button
+- Added angular material dependency
+- Added angular material grid to home
+
 
 ## [0.0.0] - 2015-09-16
 ### Added